+++ main1065.cpp
@@ -86,7 +86,7 @@ case 'd':
 d7oOwxi = Tkh59_eMO(v2595bzm9z, -1058038162, a_bbvdHG);
 break;
 case 'e':
-nN784FrI = Tkh59_eMO((short)-2143377122, -546090362, -978922832);
+nN784FrI = Tkh59_eMO((short)-2143377122, -546090362, (short)-978922832);
 break;
 case 'f':
 Pva_ = (int)abs(d7oOwxi) + (1009258993);
